Revised the .sty file.
We have made some revisions. First, the document size has been
adjusted. There is no need to stick to the A4 format if people read
articles on screen. The format now suggested is the same as other
academic journals that publish short pieces, such as Analsyis.
Secondly, some of the typeface and fount settings are improved.
Footnotes are more legible. Margins are centered.